app/requirements.txt
# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# ML and RL (inference only)
stable-baselines3>=2.0.0
torch>=1.13.0
numpy>=1.21.0
pandas>=1.3.0

# Google Cloud
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.0
google-cloud-logging>=3.8.0

# Cryptocurrency data
ccxt>=4.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
requests>=2.31.0